Conduifulate (#650)
* clap-first / Conduification

- Add an option to do a "clap-first" build
- move to CMakeRC from juce binary
- upgrade a vrariety of subordinate libraries
* State and GUI hooked up
* Note events make it play able
* VST3 for clap first working
* AUv2; Standalone; Associted Fixes; Azure dev builds it
* azure
* more assure
* upgrade juce

+++ src/engine/patch.h
@@ -159,7 +159,7 @@ struct Patch : MoveableOnly<Patch>, SampleRateSupport
     const std::unique_ptr<Part> &getPart(const PartID &pid) const
     {
         auto idx = getPartIndex(pid);
-        if (idx < 0 || idx >= parts.size())
+        if (idx < 0 || idx >= (int)parts.size())
             throw SCXTError("Unable to locate part " + pid.to_string() + " in patch " +
                             id.to_string());
         return parts[idx];
